+++ library/src/main/java/com/heartlandpaymentsystems/library/terminals/IDevice.java
@@ -5,4 +5,8 @@ import com.tsys.payments.library.domain.TransactionRequest;
 public interface IDevice {
     void doTransaction(TransactionRequest transactionRequest);
     void cancelTransaction();
+    void uploadSAF();
+    boolean isForcedSafEnabled();
+    void setForcedSafEnabled(boolean forcedSaf);
+    void acknowledgeSAFTransaction(String uniqueSafId);
 }